Double cube added and tested
Added the double cube, which creates two cubes with the grow command
instead of one.  Tested and the new action calling system works.  The
player script can trigger the grow & shrink actions on any cube that
inherits from the abstract cube class without haveing to know the
specific script it is triggering.