Limited cube growth implemented
The player now has a limited number of "charges" they use to grow cubes.
Growing a cube uses a charge, while destroying a cube gives back a
charge. A GUI element displays in the top right, telling the player how
many charges they have.

I also removed the DoubleCubeGrow script, since it was heavily outdated
and wasn't being used.  It was put in as a test/proof of concept script,
but is no longer needed.