Level work started & "cube resetter" added
Started building level 1 (scene 1 not deleted, just renamed to level 1),
which will possibly be the only level built before the code freeze.

Also added a "cube resetter" with a bland particle effect.  Touching it
removes all player created cubes and resets the number of cube charges
to max.  This will allow for a longer level.