Finished level & added win ability
Finished building level for demonstration of game.  Performance issues
will need to be addressed if my machine is to be used for demoing the
game.
Also added the ability to win.  At the end of the level, a trigger with
a blue particle effect sends the game to the main menu.  No win message
yet, but will need to add one to avoid confusion.